feat: Add InvoiceAddItemPage for selecting invoice items
- Created InvoiceAddItemPage.xaml for the UI of selecting invoice items
- Implemented InvoiceAddItemPage.xaml.cs for handling interactions and navigation
- Integrated with InvoiceAddItemViewModel for data binding and logic
