+++ POSSystem/Views/InvoiceAddItemPage.xaml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Page
-    x:Class="POSSystem.Views.InvoiceAddProductPage"
+    x:Class="POSSystem.Views.InvoiceAddItemPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:local="using:POSSystem.Views"
@@ -9,18 +9,18 @@
     mc:Ignorable="d"
     xmlns:viewmodels="using:POSSystem.ViewModels"
     Background="{ThemeResource ApplicationPageBackgroundThemeBrush}"
-    d:DataContext="{d:DesignInstance Type=viewmodels:InvoiceAddProductViewModel}">
+    d:DataContext="{d:DesignInstance Type=viewmodels:InvoiceAddItemViewModel}">
 
     <Grid Padding="30" ColumnDefinitions="1*,1*" RowDefinitions="Auto,Auto,Auto,Auto,Auto,Auto">
-        <TextBlock Text="Add Product" FontSize="30" FontFamily="{StaticResource PoppinsBold}" Grid.ColumnSpan="2" Grid.Row="0" Margin="0,0,0,30" 
+        <TextBlock Text="Add Item" FontSize="30" FontFamily="{StaticResource PoppinsBold}" Grid.ColumnSpan="2" Grid.Row="0" Margin="0,0,0,30" 
             HorizontalAlignment="Left"/>
 
         <StackPanel Grid.Row="1" Grid.ColumnSpan="2" Margin="0,0,0,20">
-            <TextBlock Text="Product Name" FontSize="14" Margin="0,0,0,8" FontFamily="{StaticResource PoppinsLight}" />
+            <TextBlock Text="Item Name" FontSize="14" Margin="0,0,0,8" FontFamily="{StaticResource PoppinsLight}" />
             <ComboBox Grid.Row="1" ItemsSource="{Binding Products, Mode=TwoWay}" 
                       FontSize="16" Padding="10" HorizontalAlignment="Stretch"
-                      SelectedItem="{Binding CurrentProduct.Name, Mode=TwoWay}" 
-                      DisplayMemberPath="Name"/>
+                      SelectedItem="{Binding CurrentProduct, Mode=TwoWay}" 
+                      DisplayMemberPath="Name" PlaceholderText="Item"/>
         </StackPanel>
 
         <StackPanel Grid.Row="2" Grid.Column="0" Margin="0,0,10,20">
@@ -30,7 +30,7 @@
 
         <StackPanel Grid.Row="2" Grid.Column="1" Margin="0,0,0,20">
             <TextBlock Text="Quantity*" FontSize="14" Margin="0,0,0,8" FontFamily="{StaticResource PoppinsLight}" />
-            <TextBox Text="{Binding CurrentProduct.Stock, Mode=TwoWay, Converter={StaticResource IntToStringConverter}}" PlaceholderText="0" FontSize="16" Padding="10"/>
+            <TextBox Text="{Binding InvoiceItem.Quantity, Mode=TwoWay}" PlaceholderText="0" FontSize="16" Padding="10"/>
         </StackPanel>
 
         <StackPanel Grid.Row="3" Grid.Column="0" Margin="0,0,20,20">
@@ -46,7 +46,7 @@
         <StackPanel Grid.Row="4" Grid.ColumnSpan="2" Orientation="Horizontal" HorizontalAlignment="Right" Margin="0,30,0,0">
             <Button Content="Discard" Click="Discard_Click"
              Width="120" Margin="0,0,10,0" Padding="10,5" FontSize="16"/>
-            <Button Content="Add" Click="Add_Click" Background="{StaticResource PinkBrush}"
+            <Button Content="Save" Click="Save_Click" Background="{StaticResource PinkBrush}"
              Width="120" Padding="10,5" FontSize="16" Foreground="{StaticResource WhiteColor}"/>
         </StackPanel>
     </Grid>
