--- POSSystem/Views/InvoiceAddProductPage.xaml.cs
@@ -1,51 +1,0 @@
-using Microsoft.UI.Xaml;
-using Microsoft.UI.Xaml.Controls;
-using Microsoft.UI.Xaml.Controls.Primitives;
-using Microsoft.UI.Xaml.Data;
-using Microsoft.UI.Xaml.Input;
-using Microsoft.UI.Xaml.Media;
-using Microsoft.UI.Xaml.Navigation;
-using POSSystem.ViewModels;
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Runtime.InteropServices.WindowsRuntime;
-using Windows.Foundation;
-using Windows.Foundation.Collections;
-
-// To learn more about WinUI, the WinUI project structure,
-// and more about our project templates, see: http://aka.ms/winui-project-info.
-
-namespace POSSystem.Views
-{
-    /// <summary>
-    /// An empty page that can be used on its own or navigated to within a Frame.
-    /// </summary>
-    public sealed partial class InvoiceAddProductPage : Page
-    {
-        private InvoiceAddProductViewModel ViewModel { get; set; }
-        public InvoiceAddProductPage()
-        {
-            ViewModel = new InvoiceAddProductViewModel();
-            this.InitializeComponent();
-            this.DataContext = ViewModel;
-        }
-
-        private void Discard_Click(object sender, RoutedEventArgs e)
-        {
-            Frame.GoBack();
-        }
-
-        private async void Add_Click(object sender, RoutedEventArgs e)
-        {
-            var dialog = new ContentDialog()
-            {
-                Title = "Add product clicked",
-                PrimaryButtonText = "OK",
-                XamlRoot = this.XamlRoot
-            };
-            await dialog.ShowAsync();
-        }
-    }
-}
